test GPU

+++ NeuralNetApproach/NCO_main.py
@@ -367,8 +367,8 @@ class NCO_main(nn.Module):
             if self.flag_DeepR:
                 if it%self.DeepR_freq==0 and it>0:
                     #print('****************Rewiring Network*********************')
-                    negative_m = torch.cat((self.W_env_to_message>0,self.W_message_to_message>0),dim=0).type(torch.FloatTensor) 
-                    negative_a = torch.cat((self.W_env_to_action>0,self.W_message_to_action>0),dim=0).type(torch.FloatTensor) 
+                    negative_m = torch.cat((self.W_env_to_message>0,self.W_message_to_message>0),dim=0)
+                    negative_a = torch.cat((self.W_env_to_action>0,self.W_message_to_action>0),dim=0)
                     negative = torch.cat( (negative_m,negative_a),dim=1 )                    
                     self.network = negative * self.network
                     for i in range(self.num_agent):
